still having errors, but got rid of error in channel enum and separated the developemnt data from the app's main
