+++ release-calendar/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "module": "commonjs",
     "esModuleInterop": true,
